backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
joblib
werkzeug
flask-jwt-extended
python-dotenv
pandas
scikit-learn

gunicorn==21.2.0

Requests==2.31.0
